Now able to get all data from firebase on start

+++ variables/structs.go
@@ -30,5 +30,5 @@ type WebhookRegistration struct {
 	Weebhook_ID string `json:"webhook_id"`
 	Url         string `json:"url"`
 	Country     string `json:"country"`
-	Calls       int    `json:"calls"`
+	Calls       int64  `json:"calls"`
 }
